// multicloudDb/resources/interfaces.ts
export interface AwsVpcResources {
  vpc: AwsVpc;
  subnets: any[] | { id: string }[];
  subnetsByName:
    | Record<string, any>
    | Record<string, { id: string; name: string }>;
  securityGroups: SecurityGroup[] | { id: string; name: string }[];
  securityGroupsByName?: Record<string, SecurityGroup>;
  securityGroupMapping: { [key: string]: Token };
  publicRouteTable: any | { id: string; name: string };
  privateRouteTable: any | { id: string; name: string };
  ec2InstanceConnectEndpoint?: Ec2InstanceConnectEndpoint;
}

// Google Cloud VPC resources interface
export interface GoogleVpcResources {
  vpc: GoogleVpc;
  subnets: ComputeSubnetwork[];
  ingressrules: ComputeFirewall[];
  egressrules: ComputeFirewall[];
  vpcLabels?: { [key: string]: string };
}

// Azure Virtual Network resources interface
export interface AzureVnetResources {
  vnet: VirtualNetwork | { name: string };
  nsgs?: { [key: string]: NetworkSecurityGroup };
  nsgRules?: { [key: string]: NetworkSecurityRule[] };
  subnets:
    | Record<string, AzureSubnet>
    | Record<string, { id: string; name: string }>;
  subnetAssociations?: SubnetNetworkSecurityGroupAssociation[];
  params?: any;
  vnetTags?: { [key: string]: string };
}